Implemented Transaction functionality, including viewing history

+++ app/src/main/java/com/budgetmanagementsystem/Goal.java
@@ -3,7 +3,7 @@ package com.budgetmanagementsystem;
 import java.util.Date;
 
 public class Goal {
-    protected int UserID;
+    protected long UserID;
     protected Date StartDate;
     protected Date EndDate;
     protected double StartBalance;
