+++ app/src/main/java/com/budgetmanagementsystem/Transaction.java
@@ -3,8 +3,8 @@ package com.budgetmanagementsystem;
 import java.util.Date;
 
 public class Transaction {
-    protected int TransactionID;
-    protected int UserID;
+    protected long TransactionID;
+    protected long UserID;
     protected String TransactionName;
     protected double TransactionAmount;
     protected Date TransactionDate;
